pyproject.toml
[tool.poetry]
name = "panthota"
version = "0.1.0"
description = "python application to manage personnal book libraries"
authors = ["Rémi DUPUY <dupuyremi@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
flask-login = "^0.6.3"
flask-sqlalchemy = "^3.1.1"
python-dotenv = "^1.0.1"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
